feat(ehr): put signature verification in core
Ref: #1040
